tui/requirements.txt
textual
httpx
websockets
pydantic
PyYAML
